requirements.txt
numpy==1.19.5
Flask==1.1.2
eventlet==0.30.2
Flask_SocketIO==5.0.1
Pillow==8.3.0
